+++ libs/pen-parser/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @protolabsai/pen-parser
 
+## 0.53.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.67.0
+
+## 0.53.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.66.0
+
 ## 0.53.12
 
 ### Patch Changes